refactor(test): consolidate SSE unit test files
Merge tests/unit/test_sse_limits.py and tests/unit/test_sse_serialization.py into tests/unit/test_sse.py to simplify the test suite. Convert the serialization test case to a standard async pytest format. Delete the redundant files.

TAG=agy
CONV=4db20f37-0e22-4128-983c-20e382b47363
